requirements.txt
ipython
google-generativeai
fastapi
jinja2
python-dotenv
